IDs in model. Added delete items in kitchen view
Added menuItemID and orderID to model.

Added functionality to remove an item from the open orders table in the KitchenView.
Not yet connected to database but this will not affect the persistent data only the copy of the data kept by the KitchenView. Still required is messaging to tell the counterView order has been prepared and is now open for payment.